move util to manifest.ts
